error files refactoring
