+++ eval/src/expr/object_constructor.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use parser::ast::Expression;
 
-use crate::{context::Context, error::EvalError, value::JmlValue};
+use crate::{context::Context, errors::EvalError, value::JmlValue};
 
 use super::eval_expr;
