+++ lexer/src/token.rs
@@ -1,9 +1,7 @@
+use crate::errors::LexingError;
 use derive_more::derive::Display;
-use errors::LexingError;
 use literal::{float, ident, int, string};
 use logos::Logos;
-
-pub mod errors;
 mod literal;
 
 #[derive(Debug, Logos, PartialEq, Clone, Display)]
